Replaces TokenSink::End with yield_end.
It was easy to forget to call yield_token(End*) before calling
end. This helps ensuring you don't. It's now symmetrical in that
yield_{start,end} both forward the token to the subsink's
yield_token. This way, the subsink can do well-formedness checking,
and the parent gets to consume the subsink, in a single call from the
source side.

Also removes the bool from yield_token. This was added before
expect_tokens, and wasn't really needed since start/end tokens are
matched.