Removes TokenSink::yield_{start,end}.
The Subsink GAT means TokenSink is not object safe. This is a problem
for generating FromTokens for types, since we can't enumerate all
subsinks (especially not for recursive data types. Attempting to make
a dyn wrapper failed, because Any: 'static, and the yield_start
returns a subsink that is 'b: 'static. This property of the parent
being able to guarantee the subsink doesn't outlive itself is sane, so
I didn't want to compromise on that. The new code does that
implicitly, as the parent owns the subsink internally.

The code becomes simpler, but as sinks are now responsible for
maintaining nesting state, the drawback is that tokens need to travel
through yield_token of every internal sink on the way to the
leaf. Whether this matters in practice is an open question that will
be answered by benchmarks.